+++ Get-Extension.ps1
@@ -252,6 +252,7 @@
                     $ExecutionContext.SessionState.InvokeCommand.GetCommand($in, 'Function,ExternalScript,Application')
                 }
 
+            #region .GetExtendedCommands
             $extCmd.PSObject.Methods.Add([psscriptmethod]::new('GetExtendedCommands', {
                 param([Management.Automation.CommandInfo[]]$CommandList)
                 $extendedCommandNames = @(
@@ -287,7 +288,8 @@
 
                 $this | Add-Member NoteProperty Extends $extends.Keys -Force
                 $this | Add-Member NoteProperty ExtensionCommands $extends.Values -Force
-            }))
+            }), $true)
+            #endregion .GetExtendedCommands
 
             if (-not $script:AllCommands) {
                 $script:AllCommands = $ExecutionContext.SessionState.InvokeCommand.GetCommands('*','Function,Alias,Cmdlet', $true)
@@ -298,6 +300,7 @@
 
             $inheritanceLevel = [ComponentModel.InheritanceLevel]::Inherited
 
+            #region .BlockComments
             $extCmd.PSObject.Properties.Add([psscriptproperty]::New('BlockComments', {
                 [Regex]::New("                   
                 \<\# # The opening tag
@@ -306,8 +309,10 @@
                 )
                 \#\> # the closing tag
                 ", 'IgnoreCase,IgnorePatternWhitespace', '00:00:01').Matches($this.ScriptBlock)
-            }))
+            }), $true)
+            #endregion .BlockComments
 
+            #region .GetHelpField
             $extCmd.PSObject.Methods.Add([psscriptmethod]::New('GetHelpField', {
                 param([Parameter(Mandatory)]$Field)
                 $fieldNames = 'synopsis','description','link','example','inputs', 'outputs', 'parameter', 'notes'
@@ -328,17 +333,26 @@
                         $match.Groups["Content"].Value -replace '[\s\r\n]+$'
                     }                    
                 }
-            }))
+            }), $true)
+            #endregion .GetHelpField
 
-            $extCmd.PSObject.Properties.Add([PSNoteProperty]::new('InheritanceLevel', $inheritanceLevel))
+            #region .InheritanceLevel
+            $extCmd.PSObject.Properties.Add([PSNoteProperty]::new('InheritanceLevel', $inheritanceLevel), $true)
+            #endregion .InheritanceLevel
+
+            #region .DisplayName
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
                 'DisplayName', [ScriptBlock]::Create("`$this.Name -replace '$extensionFullRegex'")
-            ))
+            ), $true)
+            #endregion .DisplayName
+            
+            #region .Attributes
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
                 'Attributes', {$this.ScriptBlock.Attributes}
-            ))
+            ), $true)
+            #endregion .Attributes
 
-
+            #region .Category
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
                 'Category', {
                     foreach ($attr in $this.ScriptBlock.Attributes) {
@@ -352,8 +366,10 @@
                     }
                     
                 }
-            ))
+            ), $true)
+            #endregion .Category
 
+            #region .Rank
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
                 'Rank', {
                     foreach ($attr in $this.ScriptBlock.Attributes) {
@@ -364,8 +380,10 @@
                     }
                     return 0
                 }
-            ))
+            ), $true)
+            #endregion .Rank
             
+            #region .Metadata
             $extCmd.PSObject.Properties.Add([psscriptproperty]::new(
                 'Metadata', {
                     $Metadata = [Ordered]@{}
@@ -380,21 +398,32 @@
                     }
                     return $Metadata
                 }
-            ))
+            ), $true)
+            #endregion .Metadata
 
+            #region .Description
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
                 'Description', { @($this.GetHelpField("Description"))[0] -replace '^\s+' }
-            ))
+            ), $true)
+            #endregion .Description
 
+            #region .Synopsis
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
-                'Synopsis', { @($this.GetHelpField("Synopsis"))[0] -replace '^\s+' }))
+                'Synopsis', { @($this.GetHelpField("Synopsis"))[0] -replace '^\s+' }), $true)
+            #endregion .Synopsis
 
+            #region .Examples
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
-                'Examples', { $this.GetHelpField("Example") }))
+                'Examples', { $this.GetHelpField("Example") }), $true)
+            #endregion .Examples
 
+            #region .Links
             $extCmd.PSObject.Properties.Add([PSScriptProperty]::new(
-                'Links', { $this.GetHelpField("Link") }))
+                'Links', { $this.GetHelpField("Link") }), $true
+            )
+            #endregion .Links
 
+            #region .Validate
             $extCmd.PSObject.Methods.Add([psscriptmethod]::new('Validate', {
                 param(
                     # input being validated
@@ -484,8 +513,25 @@
                 } else {
                     return $false
                 }
-            }))
+            }), $true)
+            #endregion .Validate
 
+            #region .HasValidation            
+            $extCmd.PSObject.Properties.Add([psscriptproperty]::new('HasValidation', {
+                foreach ($attr in $this.ScriptBlock.Attributes) {
+                    if ($attr -is [Management.Automation.ValidateScriptAttribute] -or
+                        $attr -is [Management.Automation.ValidateSetAttribute] -or 
+                        $attr -is [Management.Automation.ValidatePatternAttribute] -or 
+                        $attr -is [Management.Automation.ValidateRangeAttribute]) {
+                        return $true                        
+                    }
+                }
+
+                return $false
+            }), $true)            
+            #endregion .HasValidation
+
+            #region .GetDynamicParameters
             $extCmd.PSObject.Methods.Add([PSScriptMethod]::new('GetDynamicParameters', {
                 param(
                 [string]
@@ -581,9 +627,11 @@
 
                 $ExtensionDynamicParameters
 
-            }))
+            }), $true)
+            #endregion .GetDynamicParameters
 
 
+            #region .IsParameterValid
             $extCmd.PSObject.Methods.Add([PSScriptMethod]::new('IsParameterValid', {
                 param([Parameter(Mandatory)]$ParameterName, [PSObject]$Value)
 
@@ -615,8 +663,10 @@
                     }
                 }
                 return $true
-            }))
+            }), $true)
+            #endregion .IsParameterValid
 
+            #region .CouldPipe
             $extCmd.PSObject.Methods.Add([PSScriptMethod]::new('CouldPipe', {
                 param([PSObject]$InputObject)
 
@@ -663,8 +713,10 @@
                         return $mappedParams
                     }
                 }
-            }))            
+            }), $true)
+            #endregion .CouldPipe
 
+            #region .CouldRun
             $extCmd.PSObject.Methods.Add([PSScriptMethod]::new('CouldRun', {
                 param([Collections.IDictionary]$params, [string]$ParameterSetName)
 
@@ -703,7 +755,8 @@
                     return $mappedParams
                 }
                 return $false
-            }))
+            }), $true)
+            #endregion .CouldRun
 
             $extCmd.pstypenames.clear()
             if ($ExtensionTypeName) {